add proper testing
