+++ .travis.yml
@@ -20,6 +20,6 @@ install:
   - python setup.py install
 # command to run tests + check syntax style
 script:
-  - python test/test_load_module.py -v
   - flake8 examples/ ot/ test/
+  - python -m py.test -v
   # - py.test ot test
